Moved implementation details from shader.h to shader.cpp
Also improved comments in shader.h and shader.cpp clarifying how these
files differ from the originals in Vulkan-ValidationLayers.
