--- layer_factory/vkviz/render_pass.h
@@ -1,1 +1,0 @@
-/*class VkVizRenderPass {};*/
